feat: Reorganize cloud_worker better
Use more of traits to structure cloud services.
