+++ openstack_tui/src/components/identity/projects.rs
@@ -20,7 +20,8 @@ use structable_derive::StructTable;
 use tokio::sync::mpsc::UnboundedSender;
 
 use crate::{
-    action::{Action, IdentityProjectFilters, Resource},
+    action::Action,
+    cloud_worker::types::{IdentityProjectFilters, Resource},
     components::{table_view::TableViewComponentBase, Component},
     config::Config,
     mode::Mode,
