+++ openstack_tui/src/components/network/networks.rs
@@ -20,7 +20,8 @@ use structable_derive::StructTable;
 use tokio::sync::mpsc::UnboundedSender;
 
 use crate::{
-    action::{Action, NetworkNetworkFilters, NetworkSubnetFilters, Resource},
+    action::Action,
+    cloud_worker::types::{NetworkNetworkFilters, NetworkSubnetFilters, Resource},
     components::{table_view::TableViewComponentBase, Component},
     config::Config,
     mode::Mode,
