+++ openstack_tui/src/components/network/subnets.rs
@@ -19,7 +19,8 @@ use serde::Deserialize;
 use structable_derive::StructTable;
 
 use crate::{
-    action::{Action, NetworkSubnetFilters, Resource},
+    action::Action,
+    cloud_worker::types::{NetworkSubnetFilters, Resource},
     components::{table_view::TableViewComponentBase, Component},
     config::Config,
     mode::Mode,
