+++ openstack_tui/src/main.rs
@@ -25,7 +25,6 @@ use crate::{
 pub mod action;
 pub mod app;
 pub mod cli;
-mod cloud_services;
 mod cloud_worker;
 pub mod components;
 pub mod config;
